refactor(utils): move xml tags extraction to new dedicated utils file
- Moved `extractTextFromTags` from fileUtils to new xmlUtils module
- Simplified file output handling by removing `write_to_output_file` utility
- Updated imports and file organization in agent_base.py
