+++ coauthor-for-vs-code/src/utils/fileUtils.ts
@@ -330,12 +330,3 @@ export function readFileBytesSync(filePath: string): Buffer {
     throw err;
   }
 }
-
-export function extractTextFromTags(
-  inputContent: string,
-  documentTag: string,
-): string {
-  const regex = new RegExp(`<${documentTag}>(.*?)<\/${documentTag}>`, 's');
-  const match = inputContent.match(regex);
-  return match ? match[1] : inputContent;
-}
